refactor: Merge common, client and server code in single crate
