--- server/src/entities/mod.rs
@@ -1,10 +1,0 @@
-use bevy::prelude::{App, Plugin};
-
-pub mod player;
-
-pub struct ServerEntitiesPlugins;
-impl Plugin for ServerEntitiesPlugins {
-    fn build(&self, app: &mut App) {
-        app.add_plugins(player::ServerPlayerPlugin);
-    }
-}
