+++ src/core/physics.rs
@@ -1,16 +1,12 @@
 use bevy::prelude::*;
-use lightyear::prelude::FixedUpdateSet;
-use serde::{Deserialize, Serialize};
 
-use crate::network::TICK_DURATION;
-
-#[derive(Component, Debug, Reflect, Clone, PartialEq, Serialize, Deserialize, Deref, DerefMut)]
+#[derive(Component, Debug, Reflect, Clone, PartialEq, Deref, DerefMut)]
 pub struct Position(pub Vec2);
 
-#[derive(Component, Debug, Reflect, Clone, PartialEq, Serialize, Deserialize, Deref, DerefMut)]
+#[derive(Component, Debug, Reflect, Clone, PartialEq, Deref, DerefMut)]
 pub struct Velocity(pub Vec2);
 
-#[derive(Component, Debug, Reflect, Clone, PartialEq, Serialize, Deserialize, Deref, DerefMut)]
+#[derive(Component, Debug, Reflect, Clone, PartialEq, Deref, DerefMut)]
 pub struct Rotation(pub f32);
 
 #[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
@@ -22,10 +18,12 @@ impl Plugin for PhysicsPlugin {
         app.register_type::<Position>()
             .register_type::<Velocity>()
             .register_type::<Rotation>()
-            .insert_resource(Time::new_with(TICK_DURATION))
-            .configure_sets(FixedUpdate, PhysicsSet.in_set(FixedUpdateSet::TickUpdate))
-            .add_systems(FixedUpdate, update_position.in_set(PhysicsSet))
-            .add_systems(Last, update_spatial_bundles.in_set(PhysicsSet));
+            .add_systems(
+                FixedUpdate,
+                (update_position, update_spatial_bundles)
+                    .chain()
+                    .in_set(PhysicsSet),
+            );
     }
 }
