+++ src/entities/mod.rs
@@ -6,7 +6,7 @@ pub mod player;
 pub struct ClientEntitiesPlugins;
 impl Plugin for ClientEntitiesPlugins {
     fn build(&self, app: &mut App) {
-        app.add_plugins(planet::ClientPlanetPlugin)
+        app.add_plugins(planet::PlanetPlugin)
             .add_plugins(player::ClientPlayerPlugin);
     }
 }
