+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 mod gradient;
+pub mod math;
 pub mod spritesheet;
 
 #[allow(unused_imports)]
